+++ sentry-types/src/protocol/mod.rs
@@ -8,3 +8,6 @@ pub const LATEST: u16 = 7;
 
 #[cfg(feature = "protocol")]
 pub use v7 as latest;
+
+mod envelope;
+mod session;